Move `asyncIf` from the action module to the utility module.
It should have been this way from the very beginning.
