+++ lib/action.js
@@ -57,21 +57,7 @@ const readActionFiles = ( actionFiles ) => {
 	return actions;
 };
 
-const asyncIf = async ( conditionFunc, ifFunc, elseFunc ) => {
-	try {
-		await conditionFunc();
-
-		return await ifFunc();
-
-	} catch {
-		if ( elseFunc ) {
-			return await elseFunc();
-		}
-	}
-};
-
 module.exports = {
-	asyncIf,
 	createAction,
 	createPreparation,
 	readActionFile,
