+++ lib/misc.js
@@ -22,7 +22,21 @@ const parseNonSpaceSeparatedList = ( str, separator = ',' ) => {
 	return str.trim().replace( /\s*,\s*/g, separator ).split( separator );
 };
 
+const asyncIf = async ( conditionFunc, ifFunc, elseFunc ) => {
+	try {
+		await conditionFunc();
+
+		return await ifFunc();
+
+	} catch {
+		if ( elseFunc ) {
+			return await elseFunc();
+		}
+	}
+};
+
 module.exports = {
+	asyncIf,
 	generateRandomString,
 	getRootUrlFromEnv,
 	parseNonSpaceSeparatedList,
